build deveop-html same as prod
