+++ examples/no-plugins/gatsby-config.js
@@ -1,5 +1,5 @@
 module.exports = {
   siteMetadata: {
     title: `No plugins`,
-  },
+  }
 }
