+++ examples/using-postcss-sass/package.json
@@ -4,7 +4,8 @@
   "description": "Example Gatsby website using the Gatsby plugin gatsby-plugin-postcss-sass",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "develop": "gatsby develop",
+    "build": "gatsby build"
   },
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "license": "MIT",
