+++ packages/gatsby/src/utils/develop-html.js
@@ -1,0 +1,43 @@
+/* @flow */
+import webpack from "webpack"
+import Promise from "bluebird"
+import fs from "fs"
+import webpackConfig from "./webpack.config"
+const { store } = require(`../redux`)
+
+const debug = require(`debug`)(`gatsby:html`)
+
+module.exports = async (program: any) => {
+  const { directory } = program
+
+  debug(`generating static HTML`)
+  // Reduce pages objects to an array of paths.
+  const pages = store.getState().pages.map(page => page.path)
+
+  // Static site generation.
+  const compilerConfig = await webpackConfig(
+    program,
+    directory,
+    `develop-html`,
+    null,
+    ['/']
+  )
+
+  return new Promise((resolve, reject) => {
+    // console.log(compilerConfig._config)
+    webpack(compilerConfig.resolve()).run((e, stats) => {
+      console.log(e)
+      if (e) {
+        return reject(e)
+      }
+      if (stats.hasErrors()) {
+        return reject(`Error: ${stats.toJson().errors}`, stats)
+      }
+
+      // Remove the temp JS bundle file built for the static-site-generator-plugin
+      fs.unlinkSync(`${directory}/public/render-page.js`)
+
+      return resolve(null, stats)
+    })
+  })
+}
